Misc schema dir is a fallback location
This change allows placing xml files in a directory outside the gem's
misc directory, or rather, to remove the dependence on having all xml
files that are to be used exist in the same directory.

If you initialize a parser with "my/schema/dir/837.xml", later searches
for xml files (such as T101.xml) will look first in "my/schema/dir" as
before, but if the file is not be found it will search in the gem's
misc directory as well.